Update to work with Steam's new review system
Steam now only counts the total and positive percentage of purchases from the store/gift activations as the official count under under the game header. However, just above the reviews there is a new box from which you can get your own personalized count depending on what parameters you pass it. I guessed the AJAX it uses on the background and I switched to using that. The function should now be faster than what it used to.

Also, added "Chrono.gg" to the list of invalid bundles since ITAD names it that way instead of "Chrono.GG" as I had included it.